Work around SIGPIPE race condition in restartAppIfNecessary on Linux.
The APIWrapper process' stdout/stderr FDs are actually pipes to the
AIR runtime. If an application exits when restartAppIfNecessary
returns true (leading to a fork in the API call and the Steam client
being started in the subprocess), the reading end of these pipes will
be closed. If the Steam client started by the API call then tries to
write anything to stdout/stderr during startup, it receives a SIGPIPE
and is terminated.

To work around this issue, we redirect stdout/stderr to /dev/null for
the child process.