Remove Adobe AIR.framework from frameworks to link against.
For some reason, this adds a libAdobeAIR32Bridge.dylib (with a
hardcoded path of /usr/local/lib/) to the dyld dependencies of
the native library - which of course cannot be satisfied.
Since the native library is explicitly linked against the
Adobe AIR_64 shared library, all symbols can still be resolved.